Remove the ansible-pull-script lockdir in a trap handler
This makes sure that the lockdir is always removed when
ansible-pull-script exits. E.g. one doesn't have to manually remove
the lockdir in case one ends the script with Ctrl-C.

There is a small change in behavior in that lockdir is removed after
updating the script, not before. But it's probably for the better
anyway, so that the next run is sure to get the latest version.