Make getting the latest ansible-pull-script.sh more robust
By downloading the latest version to a temporary name and the renaming
that over the old one we gain a few benefits: 1) If the downloading
fails partway through, we won't corrupt the script. 2) Renaming within
a filesystem is guaranteed to be atomic, and won't change the script
we're currently running.