Use rsync instead of wget for mirroring group_vars and host_vars
With the help of the rsync --delete option, this gets rid of files
that have been deleted at the source, which can otherwise cause
problems.

Also, by using rsync instead of wget we save some of our precious
cluster insternal bandwidth, since in the vast majority of cases there
are no changes.